requirements.txt
fastapi
uvicorn[standard]
google-genai
pydantic
reportlab
